requirements.txt
appdirs==1.4.3
boto==2.46.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.8.17
django-crispy-forms==1.4.0
django-registration-redux==1.1
django-storages==1.5.2
django-toolbelt==0.0.1
djangorestframework==3.6.3
gunicorn==19.7.1
MySQL-python==1.2.5
nltk==3.2.2
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.1.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
static3==0.7.0
twython==3.4.0
whitenoise==2.0.6
